requirements.txt
flask
flask-sqlalchemy
werkzeug


python-dotenv

pymysql==1.1.0
mysqlclient
